+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Accept more options on grid table's `sd-item` directive
 - Add `shield` directive, to render badges through Shields.io
+- Add `hyper` role, to render different kinds of hyperlinks
 
 ## v0.2.1 - 2023-08-08